+++ src/syntax/tokens/regexp.ts
@@ -1,14 +1,18 @@
 import { SyntaxPalette } from "../../types/colors";
 import { TextmateToken } from "../../types/tokens";
 
-export const buildRegexpTokens = (syntaxPalette: SyntaxPalette): TextmateToken[] => {
+export const buildRegexpTokens = (
+	syntaxPalette: SyntaxPalette,
+): TextmateToken[] => {
 	return [
 		{
 			name: "REGEX GREEN",
 			scope: [
 				"string.regexp",
 				"constant.regexp",
-				"keyword.operator.or.regexp",
+				"variable.other.regexp",
+				"meta.group.regexp",
+				"constant.other.character-class.set.regexp",
 			],
 			settings: {
 				foreground: syntaxPalette.green,
@@ -17,8 +21,6 @@ export const buildRegexpTokens = (syntaxPalette: SyntaxPalette): TextmateToken[]
 		{
 			name: "REGEX CYAN",
 			scope: [
-				"constant.character.escape.backslash.regexp",
-				"meta.group.regexp",
 				"punctuation.definition.group.assertion.regexp",
 				"punctuation.definition.character-class.regexp",
 				"punctuation.character.set.begin.regexp",
@@ -33,12 +35,11 @@ export const buildRegexpTokens = (syntaxPalette: SyntaxPalette): TextmateToken[]
 			name: "REGEX PURPLE",
 			scope: [
 				"constant.character.character-class.regexp",
-				"constant.other.character-class.set.regexp",
 				"constant.other.character-class.regexp",
 				"constant.character.set.regexp",
 				"punctuation.definition.group.regexp",
 				"punctuation.definition.group.capture.regexp",
-				"keyword.control.anchor.regexp",
+				"keyword.other.back-reference.regexp",
 			],
 			settings: {
 				foreground: syntaxPalette.purple,
@@ -46,7 +47,10 @@ export const buildRegexpTokens = (syntaxPalette: SyntaxPalette): TextmateToken[]
 		},
 		{
 			name: "REGEX YELLOW",
-			scope: ["keyword.operator.quantifier.regexp"],
+			scope: [
+				"keyword.operator.or.regexp",
+				"keyword.operator.quantifier.regexp",
+			],
 			settings: {
 				foreground: syntaxPalette.yellow,
 			},
@@ -54,12 +58,26 @@ export const buildRegexpTokens = (syntaxPalette: SyntaxPalette): TextmateToken[]
 		{
 			name: "REGEX PINK",
 			scope: [
-				"keyword.operator.negation.regexp",
 				"constant.character.regexp",
 				"punctuation.definition.range.regexp",
+				"constant.character.numeric.regexp",
+				"keyword.control.anchor.regexp",
 			],
 			settings: {
 				foreground: syntaxPalette.pink,
+			},
+		},
+		{
+			name: "REGEX MAGENTA",
+			scope: [
+				"constant.character.escape.backslash.regexp",
+				"keyword.operator.negation.regexp",
+				"meta.assertion.negative-look-ahead.regexp",
+				"meta.assertion.negative-look-behind.regexp",
+				"punctuation.definition.group.no-capture.regexp",
+			],
+			settings: {
+				foreground: syntaxPalette.magenta,
 			},
 		},
 	];